+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 #### 🚀 Updates
 
 - Added a `proto unpin` command, for removing a pinned version from a `.prototools` file.
+- Updated `proto uninstall` to also entries from `.prototools` if the version was uninstalled.
 - Updated plugin locator strings to use common protocol syntax. The custom `source:` syntax is deprecated.
   - `source:./file.wasm` -> `file://./file.wasm`
   - `source:https://url.com/file.wasm` -> `https://url.com/file.wasm`